GuiYuan_Service/pyproject.toml
[project]
name = "guiyuan-service"
version = "1.0.0"
description = "归园服务端 - 在线农场×线下实地 (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "redis>=5.1.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
